refactor: inline VM_VER + always expose VM on greasyfork
