+++ scripts/webpack.conf.js
@@ -7,10 +7,13 @@ const projectConfig = require('./plaid.conf');
 const mergedConfig = shallowMerge(defaultOptions, projectConfig);
 
 const INIT_FUNC_NAME = 'VMInitInjection';
+// Copied from gulpfile.js: strip alphabetic suffix
+const VM_VER = require('../package.json').version.replace(/-[^.]*/, '');
 
 const definitions = new webpack.DefinePlugin({
   'process.env.INIT_FUNC_NAME': JSON.stringify(INIT_FUNC_NAME),
   'process.env.DEBUG': JSON.stringify(process.env.DEBUG || false),
+  'process.env.VM_VER': JSON.stringify(VM_VER),
 });
 const minimizerOptions = {
   cache: true,
